feat(api): shelves count resolver, update userbooks

+++ apps/api/libs/user-book/user-book.resolver.ts
@@ -1,13 +1,13 @@
-import { Resolver, Query, Mutation, Args } from '@nestjs/graphql';
+import { Resolver, Query, Mutation, Args, Int } from '@nestjs/graphql';
 import { UserBookService } from './user-book.service';
 import {
   BookWhereUniqueInput,
   UserBook,
-  UserBookIdentifierCompoundUniqueInput,
   UserBookUpdateInput,
+  UserBookWhereInput,
 } from '../../src/generated-db-types';
 import { AccessTokenGuard } from 'libs/auth/guards/jwt.guard';
-import { UseGuards } from '@nestjs/common';
+import { NotFoundException, UseGuards } from '@nestjs/common';
 import { CurrentUser } from 'libs/auth/decorators/currentUser.decorator';
 import { JwtPayload } from 'libs/auth/types';
 
@@ -28,6 +28,35 @@ export class UserBookResolver {
   }
 
   @UseGuards(AccessTokenGuard)
+  @Query(() => [UserBook], { nullable: true })
+  userBooks(
+    @Args('where') where: UserBookWhereInput,
+    @Args({ defaultValue: 0, name: 'offset', type: () => Int }) offset = 0,
+    @Args({ defaultValue: 20, name: 'limit', type: () => Int }) limit = 20,
+    @CurrentUser() user: JwtPayload,
+  ) {
+    return this.userBookService.findMany({
+      where,
+      userId: user.userId,
+      skip: offset,
+      take: limit,
+    });
+  }
+
+  @UseGuards(AccessTokenGuard)
+  @Query(() => Int)
+  countUserBooks(
+    @Args('where')
+    where: UserBookWhereInput,
+    @CurrentUser() user: JwtPayload,
+  ) {
+    return this.userBookService.count({
+      where,
+      userId: user.userId,
+    });
+  }
+
+  @UseGuards(AccessTokenGuard)
   @Mutation(() => UserBook)
   updateUserBook(
     @Args('data')
@@ -35,6 +64,16 @@ export class UserBookResolver {
     @Args('where') where: BookWhereUniqueInput,
     @CurrentUser() user: JwtPayload,
   ) {
+    const userBook = this.userBookService.findUnique({
+      userId: user.userId,
+      bookId: where.id,
+    });
+    if (!userBook) {
+      throw new NotFoundException(
+        `User book ${JSON.stringify(where)} does not exist`,
+      );
+    }
+
     return this.userBookService.update({
       data,
       where: {
@@ -51,6 +90,15 @@ export class UserBookResolver {
     where: BookWhereUniqueInput,
     @CurrentUser() user: JwtPayload,
   ) {
+    const userBook = this.userBookService.findUnique({
+      userId: user.userId,
+      bookId: where.id,
+    });
+    if (!userBook) {
+      throw new NotFoundException(
+        `User book ${JSON.stringify(where)} does not exist`,
+      );
+    }
     return this.userBookService.remove({
       userId: user.userId,
       bookId: where.id,
